Fixed EPR#391728

+++ include/bolt/btbb/detail/gather.inl
@@ -47,7 +47,7 @@ void gather(InputIterator1 mapfirst,
                  tbb::parallel_for (tbb::blocked_range<size_t>(0,numElements),[&](const tbb::blocked_range<size_t>& r)
                   {
                     for(size_t iter = r.begin(); iter!=r.end(); iter++)
-                        *(result + (int)iter) = * (input + mapfirst[(int)iter]); 
+                        *(result + (int)iter) = * (input + (int)mapfirst[(int)iter]); 
                   });
              }
